fix: structured, split, and modified a bunch of stuff. It’s fine. Everything is fine. Totally fine.
